npc data

+++ Core/GObject/Fighter.h
@@ -35,9 +35,9 @@ namespace GObject
 #define SKILL_ID(x) (((UInt16)(x))/SKILL_LEVEL_MAX)
 #define SKILLANDLEVEL(s,l) (((UInt16)(s))*SKILL_LEVEL_MAX | ((UInt16)(l)))
 
-#define CITTA_LEVEL(x) ((x)%CITTA_LEVEL_MAX)
-#define CITTA_ID(x) ((x)/CITTA_LEVEL_MAX)
-#define CITTAANDLEVEL(c,l) ((c)*CITTA_LEVEL_MAX | (l))
+#define CITTA_LEVEL(x) (((UInt16)(x))%CITTA_LEVEL_MAX)
+#define CITTA_ID(x) (((UInt16)(x))/CITTA_LEVEL_MAX)
+#define CITTAANDLEVEL(c,l) (((UInt16)(c))*CITTA_LEVEL_MAX | ((UInt16)(l)))
 
 #define PEERLESS_UPMAX 1
